+++ integration-test/src/test/java/org/apache/iotdb/confignode/it/regionmigration/pass/daily/datanodecrash/iotv2/batch/IoTDBRegionMigrateDataNodeCrashForIoTV2Batch.java
@@ -31,7 +31,7 @@ import org.junit.runner.RunWith;
 
 @Category({DailyIT.class})
 @RunWith(IoTDBTestRunner.class)
-public class IoTDBRegionMigrateDataNodeCrashForIoTV2IT
+public class IoTDBRegionMigrateDataNodeCrashForIoTV2Batch
     extends IoTDBRegionMigrateReliabilityITFramework {
   // region Coordinator DataNode crash tests
