+++ integration-test/src/test/java/org/apache/iotdb/confignode/it/regionmigration/pass/daily/datanodecrash/iotv2/batch/IoTDBRegionMigrateOriginalCrashWhenDeleteLocalPeerForIoTV2Batch.java
@@ -17,36 +17,28 @@
  * under the License.
  */
 
-package org.apache.iotdb.confignode.it.regionmigration.pass.commit.batch;
+package org.apache.iotdb.confignode.it.regionmigration.pass.daily.datanodecrash.iotv2.batch;
 
-import org.apache.iotdb.commons.utils.KillPoint.KillNode;
-import org.apache.iotdb.commons.utils.KillPoint.NeverTriggeredKillPoint;
-import org.apache.iotdb.confignode.it.regionmigration.IoTDBRegionMigrateReliabilityITFramework;
+import org.apache.iotdb.commons.utils.KillPoint.IoTConsensusDeleteLocalPeerKillPoints;
+import org.apache.iotdb.confignode.it.regionmigration.IoTDBRegionMigrateDataNodeCrashITFrameworkForIoTV2;
 import org.apache.iotdb.it.framework.IoTDBTestRunner;
-import org.apache.iotdb.itbase.category.ClusterIT;
+import org.apache.iotdb.itbase.category.DailyIT;
 
-import org.junit.Assert;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
+@Category({DailyIT.class})
 @RunWith(IoTDBTestRunner.class)
-@Category({ClusterIT.class})
-public class IoTDBRegionMigrateOtherIT extends IoTDBRegionMigrateReliabilityITFramework {
+public class IoTDBRegionMigrateOriginalCrashWhenDeleteLocalPeerForIoTV2Batch
+    extends IoTDBRegionMigrateDataNodeCrashITFrameworkForIoTV2 {
   @Test
-  public void badKillPoint() throws Exception {
-    try {
-      successTest(
-          1,
-          1,
-          1,
-          2,
-          buildSet(NeverTriggeredKillPoint.NEVER_TRIGGERED_KILL_POINT),
-          noKillPoints(),
-          KillNode.ALL_NODES);
-    } catch (AssertionError e) {
-      return;
-    }
-    Assert.fail("kill point not triggered but test pass");
+  public void crashBeforeDelete() throws Exception {
+    success(IoTConsensusDeleteLocalPeerKillPoints.BEFORE_DELETE);
+  }
+
+  @Test
+  public void crashAfterDelete() throws Exception {
+    success(IoTConsensusDeleteLocalPeerKillPoints.AFTER_DELETE);
   }
 }
