+++ integration-test/src/test/java/org/apache/iotdb/confignode/it/regionmigration/pass/daily/datanodecrash/iotv2/stream/IoTDBRegionMigrateCoordinatorCrashWhenRemoveRemotePeerForIoTV2Stream.java
@@ -19,7 +19,7 @@
 
 package org.apache.iotdb.confignode.it.regionmigration.pass.daily.datanodecrash.iotv2.stream;
 
-import org.apache.iotdb.commons.utils.KillPoint.IoTConsensusInactivatePeerKillPoints;
+import org.apache.iotdb.commons.utils.KillPoint.IoTConsensusRemovePeerCoordinatorKillPoints;
 import org.apache.iotdb.confignode.it.regionmigration.IoTDBRegionMigrateDataNodeCrashITFrameworkForIoTV2;
 import org.apache.iotdb.consensus.ConsensusFactory;
 import org.apache.iotdb.it.env.EnvFactory;
@@ -33,7 +33,7 @@ import org.junit.runner.RunWith;
 
 @Category({DailyIT.class})
 @RunWith(IoTDBTestRunner.class)
-public class IoTDBRegionMigrateOriginalCrashWhenRemoveRemotePeerForIoTV2IT
+public class IoTDBRegionMigrateCoordinatorCrashWhenRemoveRemotePeerForIoTV2Stream
     extends IoTDBRegionMigrateDataNodeCrashITFrameworkForIoTV2 {
 
   @Override
@@ -47,12 +47,23 @@ public class IoTDBRegionMigrateOriginalCrashWhenRemoveRemotePeerForIoTV2IT
   }
 
   @Test
-  public void crashBeforeInactivate() throws Exception {
-    success(IoTConsensusInactivatePeerKillPoints.BEFORE_INACTIVATE);
+  public void initCrash() throws Exception {
+    success(IoTConsensusRemovePeerCoordinatorKillPoints.INIT);
   }
 
   @Test
-  public void crashAfterInactivate() throws Exception {
-    success(IoTConsensusInactivatePeerKillPoints.AFTER_INACTIVATE);
+  public void crashAfterNotifyPeersToRemoveSyncLogChannel() throws Exception {
+    success(
+        IoTConsensusRemovePeerCoordinatorKillPoints.AFTER_NOTIFY_PEERS_TO_REMOVE_REPLICATE_CHANNEL);
+  }
+
+  @Test
+  public void crashAfterInactivePeer() throws Exception {
+    success(IoTConsensusRemovePeerCoordinatorKillPoints.AFTER_INACTIVE_PEER);
+  }
+
+  @Test
+  public void crashAfterFinish() throws Exception {
+    success(IoTConsensusRemovePeerCoordinatorKillPoints.FINISH);
   }
 }
