+++ integration-test/src/test/java/org/apache/iotdb/confignode/it/regionmigration/pass/daily/datanodecrash/iotv2/stream/IoTDBRegionMigrateDataNodeCrashForIoTV2Stream.java
@@ -34,7 +34,7 @@ import org.junit.runner.RunWith;
 
 @Category({DailyIT.class})
 @RunWith(IoTDBTestRunner.class)
-public class IoTDBRegionMigrateDataNodeCrashForIoTV2IT
+public class IoTDBRegionMigrateDataNodeCrashForIoTV2Stream
     extends IoTDBRegionMigrateReliabilityITFramework {
   // region Coordinator DataNode crash tests
